Changes to the user model.

+++ db/migrate/20131013072411_create_games.rb
@@ -16,12 +16,12 @@ class CreateGames < ActiveRecord::Migration
       t.integer :home_4th
       t.integer :home_score
       t.integer :home_moneyline
-      t.integer :overunder_open
-      t.integer :overunder_close
-      t.integer :spread_open
-      t.integer :spread_close
-      t.integer :overunder_secondhalf
-      t.integer :spread_secondhalf
+      t.decimal :overunder_open
+      t.decimal :overunder_close
+      t.decimal :spread_open
+      t.decimal :spread_close
+      t.decimal :overunder_secondhalf
+      t.decimal :spread_secondhalf
 
       t.timestamps
     end
